backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic>=2.5.3
pydantic-settings
alembic
argon2-cffi
google-generativeai
pytest
httpx
black
flake8
pydantic[email]
psycopg2-binary